Two bug fixes: walls placed properly. Rovers know the collection disk is 0.5 m.
The wall error was a integer division error.

+++ src/rqt_rover_gui/src/rover_gui_plugin.h
@@ -158,7 +158,7 @@ namespace rqt_rover_gui {
     map<string,int> rover_control_state;
     bool all_autonomous;
 
-    int arena_dim; // in meters
+    float arena_dim; // in meters
 
     vector<int> targets_detected;
     vector<int> targets_collected;
